requirements.txt
bcrypt==3.1.4
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
jedi==0.13.1
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.1
WTForms==2.2.1